Move _render_template_to_native_value to base class and fix numerical bounds lookup
Moves _render_template_to_native_value from QuestionOptionProcessor to the
QuestionAttributeProcessor base class so all attribute processors can use it.

Adds a NativeEnvironment-first pass to QuestionNumericalProcessor so that
complex dict-lookup templates like {{ scenario.bounds[prior.answer]['min'] }}
resolve correctly for min_value/max_value — the old _parse_template_variable
path silently skipped non-constant bracket keys and raised InvigilatorValueError.
